+++ src/core/algorithm/distribution.js
