+++ src/core/algorithm/graphToMarkovChain.js
