+++ src/core/algorithm/markovChain.js
