+++ src/core/algorithm/nodeDistribution.js
